+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.1.1 (Unreleased)
+- Add `Gecko::Record::PriceList`
 - Allow passing query parameters to `Adapter#count`
 - Add `@client.Record.first` and `@client.Record.forty_two` as helpers
 - Store the last API response at `@client.Record.last_response`